+++ scripts/cmake-build_x86.sh
@@ -162,7 +162,7 @@ defineQtVersionForCurrentDistribution() {
             "eoan")
                 releaseName='1910'
                 ;;
-            "focal")
+            "focal" | "jammy" | "noble")
                 releaseName='2004'
                 ;;
             *)
@@ -401,7 +401,7 @@ checkBoostArchive() {
     cd ${BOOST_ARCHIVE_LOCATION} || die 1 "Unable to cd into ${BOOST_ARCHIVE_LOCATION}"
     if [[ ! -e "boost_${BOOST_VERSION//./_}.tar.gz" ]]; then
         info " -> Downloading Boost archive"
-        wget https://boostorg.jfrog.io/artifactory/main/release/${BOOST_VERSION}/source/boost_${BOOST_VERSION//./_}.tar.gz
+        wget https://archives.boost.io/release/${BOOST_VERSION}/source/boost_${BOOST_VERSION//./_}.tar.gz
     else
         info " -> Using existing Boost archive"
     fi